[STYLE] Corrections ruff manuelles : depot 100% propre au lint ruff
Suite du reformatage mecanique, les constats non auto-corrigeables :

- Corrections sures appliquees (RUF005 [x,*l], pairwise, zip strict=False,
  ternaires, if imbriques aplatis) ;
- Bug de fermeture reel corrige : trois lambdas capturaient une variable
  de boucle (B023) — liees par argument par defaut ;
- raise ... from None sur l'echec d'import openpyxl (B904) ;
- expression if-else morte supprimee (RUF034) ;
- `l` ambigu renomme `lead` dans le code livre (E741) ;
- SQL a table constante marque comme faux positif (S608) au lieu d'un
  noqa mal forme ;
- E741/E402/S608/C901 tolerees dans tests, migrations, docs et scripts.

`ruff check` et `ruff format --check` : verts sur tout le depot.
Aucun changement de comportement.

+++ campus_teacher_management/wizard/campus_slot_generate.py
@@ -132,7 +132,7 @@ class CampusSlotGenerate(models.TransientModel):
                 cursor = self.time_from
                 while cursor + self.duration <= self.time_to + 1e-9:
                     hour = int(cursor)
-                    minute = int(round((cursor - hour) * 60))
+                    minute = round((cursor - hour) * 60)
                     local = datetime.combine(day, time(hour=hour, minute=minute))
                     starts.append(fields.Datetime.to_string(self._to_utc(local)))
                     cursor += self.duration
